+++ Client/src/components/StockTrading/StockTradingListTable.tsx
