+++ Client/src/components/StockTrading/StockTradingTableRow.tsx
